// server/utils/types.ts
export type JsonPropertiesFeed = EditorLayerFeed | AdStudioLayerFeed
// ...
export type JsonTransition = {
  type: JsonTransitionType
  direction: JsonTransitionDirection
  duration: number
  delay: number
  tween: JsonTweenType
  ease: JsonEaseType
  wordsDuration?: number
  wordsAppearOrder?: 'l2r' | 'r2l'
  slidePosX: number
  slidePosY: number
  alphaOffset: number
  slideOffset: number
  blurAmount: number
  zoom: number
  preset: string
  color?: string
  shadowLength?: number
  advancedProperties?: AdvancedProperties
}

export type JsonTransitionGradColor = {
  gradColors: JsonGradColor[]
}

export type JsonTransitionMid = {
  type: JsonTransitionType
  direction: 'vertical' | 'horizontal' | 'top'
  variation?: string | null
  duration: number
  delay: number
  motionTime: number
  ease: JsonEaseType
  tween: JsonTweenType
  preset: string
  startEnd: string
  advancedProperties?: AdvancedProperties
  gradient?: JsonTransitionGradColor
}
// ...
export type JsonShadow = {
  useShadow: boolean
  hShadow: number
  vShadow: number
  blur: number
  spread: number
  color: string
}

export type JsonBlur = {
  useBlur: boolean
  pixels: number
}

export const JsonFlipTypeArr = [
  'both',
  'horizontal',
  'none',
  'vertical',
] as const

export type JsonFlipType = (typeof JsonFlipTypeArr)[number]

export type JsonTextAlignment = 'left' | 'right' | 'center' | 'justify'

export const JsonTextVerticalAlignment = {
  top: 'top',
  center: 'center',
  bottom: 'bottom',
} as const
export type JsonTextVerticalAlignmentType =
  (typeof JsonTextVerticalAlignment)[keyof typeof JsonTextVerticalAlignment]
// ...
export const JsonTextContentHeight = {
  auto: 'auto',
  fixed: 'fixed',
} as const
export type JsonTextContentHeightType =
  (typeof JsonTextContentHeight)[keyof typeof JsonTextContentHeight]
// ...
export type JsonAction =
  | JsonActionGotoURL
  | JsonActionGotoSlide
  | JsonActionNone
  | JsonActionVideoEnd

export function isJsonActionGotoSlide(
  action: JsonAction
): action is JsonActionGotoSlide {
  return action.type === 'gotoSlide'
}

export function isJsonActionGotoURL(
  action: JsonAction
): action is JsonActionGotoURL {
  return action.type === 'gotoURL'
}

export function isJsonActionNone(action: JsonAction): action is JsonActionNone {
  return action.type === 'none'
}

export type JsonBaseElementProperties = {
  id: number
  bannersetElementId?: number | null
  elementCategory?: string | null
  layerName: string
  dropShadow?: JsonShadow
  blur?: JsonBlur
  opacity: number
  rotation: number
  flip?: JsonFlipType
  x: number
  y: number
  width: number
  height: number
  buildIn?: JsonTransition
  buildMid?: JsonTransitionMid
  buildOut?: JsonTransition
  actions?: JsonAction[]
  locked?: boolean
  lockedProperties?: string[]
  visible: boolean
  group?: number | null
  feed?: JsonPropertiesFeed
}

export type JsonTextProperties = JsonBaseElementProperties & {
  text?: string
  alignment: JsonTextAlignment
  verticalAlign?: JsonTextVerticalAlignmentType
  contentHeightType: JsonTextContentHeightType
  fontSize: number
  lineHeight: number
  letterSpacing: number
  config: JsonTextSlateConfig // TODO:
  textShadow?: JsonShadow
  initialFontSize?: number
  textDirection?: string | null
  scale?: number
  blendMode?: JsonBlendMode
  outline?: JsonTextOutline
}

export type JsonTextOutline = {
  useOutline?: boolean
  color?: string
  weight?: number
}

export type JsonText = {
  type: 'layer'
  layerType: 'text'
  properties: JsonTextProperties
}

export type JsonOldTextV2 = {
  type: 'layer'
  layerType: 'text'
  properties: JsonTextProperties & { config: JsonTextOldConfig[] }
}

export type JsonOldTextV1 = {
  properties: JsonTextOldConfigStyle
}

// TODO: add all attributes
export type JsonTextSlateConfigNode = {
  defaultFontSettings?: JsonFontSettings
  children: Array<JsonTextSlateConfigChildren | JsonTextSlateConfigNode>
  type:
    | 'paragraph'
    | 'legacy'
    | 'subparagraph'
    | 'wrapper'
    | 'bullet-list'
    | 'number-list'
    | 'list-item'
    | 'feed-chip'
}
// ...
export type Boundary = {
  type?: JsonTextContentBoundaryType
  minFontSize?: number
}

export type JsonTextSlateConfig = {
  type: 'slate'
  nodes: JsonTextSlateConfigNode[]
  selection?: unknown
  keepSelection?: boolean
  boundary?: Boundary
}
// ...
export type JsonBlendMode =
  | 'color'
  | 'color-burn'
  | 'color-dodge'
  | 'darken'
  | 'difference'
  | 'exclusion'
  | 'hard-light'
  | 'hue'
  | 'lighten'
  | 'luminosity'
  | 'multiply'
  | 'normal'
  | 'overlay'
  | 'saturation'
  | 'screen'
  | 'soft-light'